backend/requirements.txt
django=3.2.7
djangorestframework=3.13.1
markdown
django-filter
djangorestframework-simplejwt